pyproject.toml
[project]
name = "oraja-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "imagehash>=4.3.2",
    "ipython>=9.4.0",
    "obsws-python>=1.8.0",
    "packaging>=25.0",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
